Remove PostingWithClustersScorer
